concord-server: use ProjectLoader interface, bind explicitly
Before this change, RepositoryManager had a hard dependency on
a specific implementation of ProjectLoader and was hard coded to treat
any repository as a "project" (i.e. containing files for v1, v2 or
other supported runtimes).

The change allows simpler testing and re-use in other projects.
