+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/ProcessModule.java
@@ -23,7 +23,8 @@ package com.walmartlabs.concord.server.process;
 import com.google.inject.Binder;
 import com.google.inject.Module;
 import com.walmartlabs.concord.imports.ImportManager;
-import com.walmartlabs.concord.process.loader.ProjectLoader.ProjectLoaderConfiguration;
+import com.walmartlabs.concord.process.loader.ConcordProjectLoader;
+import com.walmartlabs.concord.process.loader.ProjectLoader;
 import com.walmartlabs.concord.server.process.checkpoint.ProcessCheckpointResource;
 import com.walmartlabs.concord.server.process.checkpoint.ProcessCheckpointV2Resource;
 import com.walmartlabs.concord.server.process.event.ProcessEventDao;
@@ -74,7 +75,8 @@ public class ProcessModule implements Module {
         binder.bind(ProcessStateManager.class).in(SINGLETON);
         binder.bind(ProcessWaitManager.class).in(SINGLETON);
 
-        binder.bind(ProjectLoaderConfiguration.class).toProvider(ProjectLoaderConfigurationProvider.class);
+        binder.bind(ConcordProjectLoader.Configuration.class).toProvider(ConcordProjectLoaderConfigurationProvider.class);
+        binder.bind(ProjectLoader.class).to(ConcordProjectLoader.class);
 
         bindSingletonScheduledTask(binder, ProcessCleaner.class);
         bindSingletonScheduledTask(binder, ProcessLocksWatchdog.class);
