+++ server/impl/src/main/java/com/walmartlabs/concord/server/repository/RepositoryManager.java
@@ -23,7 +23,7 @@ package com.walmartlabs.concord.server.repository;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.walmartlabs.concord.common.IOUtils;
 import com.walmartlabs.concord.dependencymanager.DependencyManager;
-import com.walmartlabs.concord.process.loader.ProjectLoader;
+import com.walmartlabs.concord.process.loader.ConcordProjectLoader;
 import com.walmartlabs.concord.repository.*;
 import com.walmartlabs.concord.sdk.Secret;
 import com.walmartlabs.concord.server.cfg.GitConfiguration;
@@ -107,7 +107,7 @@ public class RepositoryManager {
                             .build(), path);
 
             if (repoCfg.isConcordFileValidationEnabled()) {
-                if (!ProjectLoader.isConcordFileExists(repo.path())) {
+                if (!ConcordProjectLoader.isConcordFileExists(repo.path())) {
                     throw new InvalidRepositoryPathException("Invalid repository path: `concord.yml` or `.concord.yml` is missing!");
                 }
             }
